requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
black==24.10.0
certifi==2024.12.14
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
